Implement build with MSys2. Additions and changes: * Fixed build `mk` utility from scratch * Reworked some mkfiles - cmd-style commands replaced with bash-style ones; resolved cycle build issues. * Some tiny code modifications

+++ utils/liballoc/alloc.c
@@ -30,7 +30,7 @@ kcalloc(size_t n, size_t szelem)
 }	
 
 void*
-mallocz(size_t n, int clr)
+mallocz(ulong n, int clr)
 {
 	void *v;
 	v = malloc(n);
